# Force to variable depths based on score