--- a/testdata/3x4-2520743-1400.txt
+++ b/testdata/3x4-2520743-1400.txt
@@ -51,7 +51,7 @@ Found unbreakable board for 2520743: perslitesand
   "boards_to_test": 8078,
   "expanded_to_test": 8078,
   "init_nodes": 1960065,
-  "total_nodes": 2362580,
+  "total_nodes": 1960065,
   "freed_nodes": 0,
   "n_bound": 5
 }