Fixes #57 
Fixes #58 

Originally, `HybridTreeBreaker` would lift/force a pre-determined number of cells. #37 made it possible for this to vary based on the tree size. Now it makes a decision about when to switch based on the bound of the subtree under consideration. If it's high, we keep forcing more cells. If it's low enough (or we've forced enough cells already), then we switch over to `orderly_bound`.

This has a dramatic effect on hard board classes, particularly 4x4 board classes:

- 500 board test set (b026203) 14000s → 5500s
- best class (19005578): 4.28GB / 4438.13s → 1254s / 4.75GB

So a 2-3x speedup without too much of a RAM penalty. The effect is much smaller for 3x4 boards.

In practice, this forces way more cells than I ever let it before. It would force all 16 cells if I didn't put a cap. This wouldn't have worked before, but thanks to #57 the memory penalty is much lower for more forces than it used to be.

I don't think score is exactly the right way to determine when to switch to `orderly_bound`, but this is a big enough improvement that I don't want to sweat the details.